// src/store/store.d.ts
export interface IStoreData {
  uniq_id: string;
  product_name: string;
  manufacturer: string;
  price: string;
  number_available_in_stock: string;
  number_of_reviews: number;
  number_of_answered_questions: number;
  average_review_rating: string;
  amazon_category_and_sub_category: string;
  customers_who_bought_this_item_also_bought: string;
  description: string;
  product_information: string;
  product_description: string;
  items_customers_buy_after_viewing_this_item: string;
  customer_questions_and_answers: string;
  customer_reviews: string;
  sellers: string;
}